requirements.txt
base58==1.0.3
crypto==1.4.1
cryptography==2.6.1
ecdsa==0.13
grpcio==1.21.1
grpcio-tools==1.21.1
ledgerblue==0.1.23
protobuf==3.7.1
secp256k1==0.13.2
eth-keys==0.2.1